Update on Distance determination
Update For Oriented Normals.

By using the result of the dot product rather than the previous "tick" system allows for a more qualitive comparison that does not rely on a predictable average displacement from neighboring terrain elements

Also switched from the inverse exponential coloring to a polynomial curve on the basis that it results in a marginally superior gray scale result.